improve sampling stability, HID WIP

+++ include/ResistorLadderButtons.hpp
@@ -1,6 +1,6 @@
 class ResistorLadderButtons {
-	public:
-	ResistorLadderButtons(const uint16 adcMax, const uint16 rBase, const uint8 pin, const uint16 * bins, const uint8 binCount);
+  public:
+	ResistorLadderButtons(const uint16 adcMax, const uint16 rBase, const uint8 pin, const uint32* bins, const uint8 binCount);
 	void setup();
 	void loop();
 
@@ -8,17 +8,17 @@ class ResistorLadderButtons {
 	void setButtonDownCallback(void (*callback)(const uint8 button)) { buttonDownCallback = callback; }
 	void setButtonUpCallback(void (*callback)(const uint8 button, uint32 duration)) { buttonUpCallback = callback; }
 	void setButtonHeldCallback(void (*callback)(const uint8 button, uint32 duration)) { buttonHeldCallback = callback; }
-	void setDebugStream(Stream * stream) { debugStream = stream; }
+	void setDebugStream(Stream* stream) { debugStream = stream; }
 
-	private:
-	int8 rToBin(uint16 r);
-	uint16 adcToResistance(uint16 rBase, uint16 adc);
+  private:
+	int8 rToBin(uint32 r);
+	uint32 adcToResistance(uint16 rBase, uint16 adc);
 
 	// fields
 	const uint16 adcMax;
 	const uint16 rBase;
 	const uint8 pin;
-	const uint16 * bins;
+	const uint32* bins;
 	const uint8 binCount;
 
 	int16 lastBin = -1;
@@ -31,5 +31,5 @@ class ResistorLadderButtons {
 	void (*buttonDownCallback)(const uint8 button);
 	void (*buttonUpCallback)(const uint8 button, uint32 duration);
 	void (*buttonHeldCallback)(const uint8 button, uint32 duration);
-	Stream * debugStream = NULL;
+	Stream* debugStream = NULL;
 };
